Extract Composer helpers into separate files
